[WIP] Refactoring & Setup Personal-Details-Form
