+++ src/app/(main)/(routes)/onboarding/personal-details/page.tsx
@@ -1,9 +1,8 @@
 import ProtectedSiteHeader from "@/app/_components/nav/protected-site-header";
-import { OnboardingForm } from "@/app/_components/forms/onboarding-form";
 import { type Metadata } from "next";
 
 export const metadata: Metadata = {
-  title: "Onboarding | HackPSH",
+  title: "Onboarding | Personal Details | HackPSH",
   description: "Weclome to HackPSH!",
 }
 
@@ -14,7 +13,7 @@ export default function Page() {
     <>
       <ProtectedSiteHeader />
       <div className="container relative flex-col items-center justify-center space-y-6">
-        <OnboardingForm />
+        Personal Details Form
       </div>
     </>
   )
