+++ src/app/_lib/zod-schemas/forms/onboarding/support-us.tsx
@@ -1,0 +1,8 @@
+import { z } from "zod";
+
+export const SupportUsFormSchema = z.object({
+  be_officer: z.boolean(),
+  type_officer: z.enum(["administrative", "technical", "both"]),
+});
+
+export type THelpForm = z.infer<typeof SupportUsFormSchema>;
